requirements.txt
torch==2.2.0+cu118
torchvision==0.17.0+cu118
torchaudio==2.2.0
numpy==1.25.2
Pillow==10.1.0
opencv-python==4.9.0.72
tqdm==4.66.1
scikit-learn==1.3.2
matplotlib==3.8.1
